MAINT: no need to cast in asarray

+++ scipy/stats/stats.py
@@ -557,7 +557,7 @@ def tmean(a, limits=None, inclusive=(True, True), axis=None):
     10.0
 
     """
-    a = asarray(a, dtype=np.float64)
+    a = asarray(a)
     if limits is None:
         return np.mean(a, axis)
     am = _mask_to_limits(a, limits, inclusive)
